Use the css var for real feedback
This also applies the css var change to html element instead
of hardcoded classes on any element so that we only need to click
on the document root to check whatever information we need in the
developer console.

We can now optionally render the value of the screen-size css var
on the page itself by applying the class 'report' to the html
element.